Improved Actionable/Processor pattern, improved use of generics and more
- Added IntAether4DAsymmetricSectionSwap and related classes
- Removed classes that become unnecessary due to improved use of
generics
- Added ActionableGridTransformerProcessor class to reduce repeated code
- Other minor changes

+++ CellularAutomata/src/cellularautomata/automata/SizeLimitedAnisotropicLongGrid3DBlock.java
@@ -74,7 +74,7 @@ public class SizeLimitedAnisotropicLongGrid3DBlock implements LongGrid3D, Serial
 		int x = minX;
 		long roundedSize = Utils.roundUpToEightMultiple(size);
 		while (roundedSize <= maxBytes) {
-			size += AnisotropicLongGrid3DSlice.getSliceSize(x) + Long.BYTES;
+			size += AnisotropicLongGrid3DSlice.getSliceSize(x) + Integer.BYTES;
 			roundedSize = Utils.roundUpToEightMultiple(size);
 			x++;
 			xLength++;
